IZPACK-1340: Use Assume.assumeTrue() to skip test execution on Windows without administrative rights.

+++ izpack-test/src/test/java/com/izforge/izpack/integration/windows/WindowsConsoleInstallationTest.java
@@ -34,6 +34,7 @@ import static org.junit.Assert.assertTrue;
 import java.io.File;
 
 import org.junit.After;
+import org.junit.Assume;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -81,6 +82,8 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
 {
     private final boolean skipTests = new PrivilegedRunner(Platforms.WINDOWS).isElevationNeeded();
 
+    private final boolean isAdminUser = new PrivilegedRunner(Platforms.WINDOWS).isAdminUser();
+    
     /**
      * The installer container.
      */
@@ -151,19 +154,17 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @Override
     public void setUp() throws Exception
     {
-        if(!skipTests) {
-            //assertFalse("This test must be run as administrator, or with Windows UAC turned off",
-            //    new PrivilegedRunner(Platforms.WINDOWS).isElevationNeeded());
-            super.setUp();
-            String appName = getInstallData().getInfo().getAppName();
-            assertNotNull(appName);
-            File file = FileUtil.getLockFile(appName);
-            if (file.exists()) {
-                assertTrue(file.delete());
-            }
-
-            destroyRegistryEntries();
+    	Assume.assumeTrue("This test must be run as administrator, or with Windows UAC turned off", !skipTests && isAdminUser);
+    	    	
+        super.setUp();
+        String appName = getInstallData().getInfo().getAppName();
+        assertNotNull(appName);
+        File file = FileUtil.getLockFile(appName);
+        if (file.exists()) {
+            assertTrue(file.delete());
         }
+
+        destroyRegistryEntries();
     }
 
     /**
@@ -174,13 +175,15 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @After
     public void tearDown() throws Exception
     {
-        if(!skipTests) {
-            destroyRegistryEntries();
+    	// don't use Assume in @After methods!
+    	
+    	if ((!skipTests && isAdminUser)) {
+    		destroyRegistryEntries();
+    	}
 
-            if (getUninstallerJar() != null) {
-                // remove the uninstaller dir
-                FileUtils.deleteRecursively(getUninstallerJar().getParentFile());
-            }
+    	if (getUninstallerJar() != null) {
+            // remove the uninstaller dir
+            FileUtils.deleteRecursively(getUninstallerJar().getParentFile());
         }
     }
 
@@ -193,18 +196,18 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @InstallFile("samples/windows/install.xml")
     public void testInstallation() throws Exception
     {
-        if(!skipTests) {
-            // run the install
-            checkInstall(container, APP_NAME);
-            assertTrue(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
+    	Assume.assumeTrue("This test must be run as administrator, or with Windows UAC turned off", !skipTests && isAdminUser);
 
-            // run the uninstaller and verify that uninstall key is removed
-            File uninstaller = getUninstallerJar();
-            assertTrue(uninstaller.exists());
-            UninstallHelper.consoleUninstall(uninstaller);
+    	// run the install
+        checkInstall(container, APP_NAME);
+        assertTrue(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
 
-            assertFalse(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
-        }
+        // run the uninstaller and verify that uninstall key is removed
+        File uninstaller = getUninstallerJar();
+        assertTrue(uninstaller.exists());
+        UninstallHelper.consoleUninstall(uninstaller);
+
+        assertFalse(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
     }
 
     /**
@@ -216,33 +219,33 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @InstallFile("samples/windows/install.xml")
     public void testMultipleInstallation() throws Exception
     {
-        if(!skipTests) {
-            // run the install
-            checkInstall(container, APP_NAME);
+    	Assume.assumeTrue("This test must be run as administrator, or with Windows UAC turned off", !skipTests && isAdminUser);
+    	
+        // run the install
+        checkInstall(container, APP_NAME);
 
-            // remove the lock file to enable second installation
-            removeLock();
+        // remove the lock file to enable second installation
+        removeLock();
 
-            // run the installation again
-            ConsoleInstallerContainer container2 = new TestConsoleInstallerContainer();
-            TestConsoleInstaller installer2 = container2.getComponent(TestConsoleInstaller.class);
-            InstallData installData2 = container2.getComponent(InstallData.class);
-            TestConsole console2 = installer2.getConsole();
-            console2.addScript("CheckedHelloPanel", "y", "1");
-            console2.addScript("TargetPanel", "Y", "\n", "1");
-            console2.addScript("PacksPanel", "1");
-            console2.addScript("ShortcutPanel", "N", "N", "1");
-            console2.addScript("SimpleFinishPanel", "1");
+        // run the installation again
+        ConsoleInstallerContainer container2 = new TestConsoleInstallerContainer();
+        TestConsoleInstaller installer2 = container2.getComponent(TestConsoleInstaller.class);
+        InstallData installData2 = container2.getComponent(InstallData.class);
+        TestConsole console2 = installer2.getConsole();
+        console2.addScript("CheckedHelloPanel", "y", "1");
+        console2.addScript("TargetPanel", "Y", "\n", "1");
+        console2.addScript("PacksPanel", "1");
+        console2.addScript("ShortcutPanel", "N", "N", "1");
+        console2.addScript("SimpleFinishPanel", "1");
 
-            assertFalse(registryKeyExists(handler, UNINSTALL_KEY2));
-            checkInstall(installer2, installData2);
+        assertFalse(registryKeyExists(handler, UNINSTALL_KEY2));
+        checkInstall(installer2, installData2);
 
-            // verify the UNINSTALL_NAME has been updated
-            assertEquals(APP_NAME + "(1)", installData2.getVariable("UNINSTALL_NAME"));
+        // verify the UNINSTALL_NAME has been updated
+        assertEquals(APP_NAME + "(1)", installData2.getVariable("UNINSTALL_NAME"));
 
-            // verify a second key is created
-            assertTrue(registryKeyExists(handler, UNINSTALL_KEY2));
-        }
+        // verify a second key is created
+        assertTrue(registryKeyExists(handler, UNINSTALL_KEY2));
     }
 
     /**
@@ -254,32 +257,32 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @InstallFile("samples/windows/install.xml")
     public void testRejectMultipleInstallation() throws Exception
     {
-        if(!skipTests) {
-            checkInstall(container, APP_NAME);
+    	Assume.assumeTrue("This test must be run as administrator, or with Windows UAC turned off", !skipTests && isAdminUser);
+    	
+        checkInstall(container, APP_NAME);
 
-            removeLock();
+        removeLock();
 
-            ConsoleInstallerContainer container2 = new TestConsoleInstallerContainer();
-            TestConsoleInstaller installer2 = container2.getComponent(TestConsoleInstaller.class);
-            RegistryDefaultHandler handler2 = container2.getComponent(RegistryDefaultHandler.class);
-            InstallData installData2 = container2.getComponent(InstallData.class);
+        ConsoleInstallerContainer container2 = new TestConsoleInstallerContainer();
+        TestConsoleInstaller installer2 = container2.getComponent(TestConsoleInstaller.class);
+        RegistryDefaultHandler handler2 = container2.getComponent(RegistryDefaultHandler.class);
+        InstallData installData2 = container2.getComponent(InstallData.class);
 
-            TestConsole console2 = installer2.getConsole();
-            console2.addScript("CheckedHelloPanel", "n");
+        TestConsole console2 = installer2.getConsole();
+        console2.addScript("CheckedHelloPanel", "n");
 
-            assertFalse(registryKeyExists(handler2, UNINSTALL_KEY2));
-            installer2.run(Installer.CONSOLE_INSTALL, null, new String[0]);
+        assertFalse(registryKeyExists(handler2, UNINSTALL_KEY2));
+        installer2.run(Installer.CONSOLE_INSTALL, null, new String[0]);
 
-            // verify the installation thinks it was unsuccessful
-            assertFalse(installData2.isInstallSuccess());
+        // verify the installation thinks it was unsuccessful
+        assertFalse(installData2.isInstallSuccess());
 
-            // make sure the script has completed
-            TestConsole console = installer2.getConsole();
-            assertTrue("Script still running panel: " + console.getScriptName(), console.scriptCompleted());
+        // make sure the script has completed
+        TestConsole console = installer2.getConsole();
+        assertTrue("Script still running panel: " + console.getScriptName(), console.scriptCompleted());
 
-            // verify the second registry key hasn't been created
-            assertFalse(registryKeyExists(handler2, UNINSTALL_KEY2));
-        }
+        // verify the second registry key hasn't been created
+        assertFalse(registryKeyExists(handler2, UNINSTALL_KEY2));
     }
 
     /**
@@ -292,30 +295,30 @@ public class WindowsConsoleInstallationTest extends AbstractConsoleInstallationT
     @InstallFile("samples/windows/consoleinstall_alt_uninstall.xml")
     public void testNonDefaultUninstaller() throws Exception
     {
-        if(!skipTests) {
-            assertFalse(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
+    	Assume.assumeTrue("This test must be run as administrator, or with Windows UAC turned off", !skipTests && isAdminUser);
+    	
+        assertFalse(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
 
-            TestConsole console = installer.getConsole();
-            console.addScript("CheckedHelloPanel", "y", "1");
-            console.addScript("InfoPanel", "1");
-            console.addScript("TargetPanel", "\n", "1");
-            console.addScript("InstallPanel", "1");
-            console.addScript("FinishPanel", "1");
+        TestConsole console = installer.getConsole();
+        console.addScript("CheckedHelloPanel", "y", "1");
+        console.addScript("InfoPanel", "1");
+        console.addScript("TargetPanel", "\n", "1");
+        console.addScript("InstallPanel", "1");
+        console.addScript("FinishPanel", "1");
 
-            //run installer and check that default uninstaller doesn't exist
-            InstallData installData = getInstallData();
-            checkInstall(installer, installData, false);
+        //run installer and check that default uninstaller doesn't exist
+        InstallData installData = getInstallData();
+        checkInstall(installer, installData, false);
 
-            //check that uninstaller exists as specified in install spec
-            String installPath = installData.getInstallPath();
-            assertTrue(new File(installPath, "/uninstallme.jar").exists());
+        //check that uninstaller exists as specified in install spec
+        String installPath = installData.getInstallPath();
+        assertTrue(new File(installPath, "/uninstallme.jar").exists());
 
-            //check that the registry key has the correct value
-            assertTrue(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
-            String command = "\"" + installData.getVariable("JAVA_HOME") + "\\bin\\javaw.exe\" -jar \"" + installPath
-                + "\\uninstallme.jar\"";
-            registryValueStringEquals(handler, DEFAULT_UNINSTALL_KEY, UNINSTALL_CMD_VALUE, command);
-        }
+        //check that the registry key has the correct value
+        assertTrue(registryKeyExists(handler, DEFAULT_UNINSTALL_KEY));
+        String command = "\"" + installData.getVariable("JAVA_HOME") + "\\bin\\javaw.exe\" -jar \"" + installPath
+            + "\\uninstallme.jar\"";
+        registryValueStringEquals(handler, DEFAULT_UNINSTALL_KEY, UNINSTALL_CMD_VALUE, command);
     }
 
     /**
